fix(analytics): round minute fields to whole minutes
Ensure analytics minute-based fields (sleep and duration) are rounded to integers in API responses and pediatrician summary displays.

+++ analytics/serializers.py
@@ -22,9 +22,9 @@ class DailyDataSerializer(serializers.Serializer):
 
     date = serializers.DateField()
     count = serializers.IntegerField()
-    average_duration = serializers.FloatField(allow_null=True)
+    average_duration = serializers.IntegerField(allow_null=True)
     total_oz = serializers.FloatField(allow_null=True)
-    total_minutes = serializers.FloatField(allow_null=True, required=False)
+    total_minutes = serializers.IntegerField(allow_null=True, required=False)
 
 
 class WeeklySummaryDataSerializer(serializers.Serializer):
